double value provided for interface, certain variables inlined, redundant setter removed, block of code drawing default figure moved out from if-else

+++ src/main/java/core/basesyntax/Circle.java
@@ -9,7 +9,7 @@ public class Circle extends Figure {
     }
 
     @Override
-    public int getArea() {
+    public double getArea() {
         return (int) (Math.PI * radius * radius);
     }
